requirements.txt
django>=2.1.4
djangorestframework>=3.9.0
djangorestframework-jwt>=1.11.0
psycopg2>=2.7.6.1
